refine Window Executor Test (#5464)
ref pingcap/tiflash#4609, close pingcap/tiflash#5465
